show deprecated operator post installation e2e tests
